Update Release-Engine-Staging from UE5/Main @21717042,23021685 #rnx
[CL 23023039 by Marc Audy in ue5-release-engine-staging branch]

+++ Engine/Source/Editor/MovieSceneTools/Private/EditModes/SkeletalAnimationTrackEditMode.cpp
@@ -308,7 +308,7 @@ void FSkeletalAnimationTrackEditMode::Render(const FSceneView* View, FViewport* 
 											UE::Anim::FStackAttributeContainer TempAttributes;
 											FAnimationPoseData AnimationPoseData(OutPose, OutCurve, TempAttributes);
 
-											const float Seconds = static_cast<float>(AnimSection->MapTimeToAnimation(CurrentTime.Time, CurrentTime.Rate));
+											const double Seconds = AnimSection->MapTimeToAnimation(CurrentTime.Time, CurrentTime.Rate);
 											FAnimExtractContext ExtractionContext(Seconds, false);
 
 											AnimSequence->GetAnimationPose(AnimationPoseData, ExtractionContext);
