+++ Engine/Source/Editor/UnrealEd/Private/FeedbackContextEditor.cpp
@@ -149,7 +149,7 @@ public:
 			.Padding(FMargin(FixedPaddingH))
 			[
 				SNew(SBox).
-				WidthOverride(FixedWidth) 
+				WidthOverride(static_cast<float>(FixedWidth))
 				[
 					VerticalBox 
 				]
@@ -234,7 +234,7 @@ private:
 			.AutoHeight()
 			[
 				SNew(SBox)
-				.HeightOverride(SecondaryBarHeight)
+				.HeightOverride(static_cast<float>(SecondaryBarHeight))
 				[
 					SNew(SProgressBar)
 					.BorderPadding(FVector2D::ZeroVector)
